+++ designsystem/src/main/java/com/juco/designsystem/dialog/BreakTimeSelectionDialog.kt
@@ -1,6 +1,5 @@
-package com.juco.common.dialog
+package com.juco.designsystem.dialog
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -20,12 +19,11 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
-import com.juco.common.BreakTimeInputTextField
+import com.juco.designsystem.inputfield.BreakTimeInputTextField
 
 @Composable
 fun BreakTimeSelectionDialog(
